requirements.txt
aiobotocore==1.1.2
aiohttp==3.7.3
aioitertools==0.7.1
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
boto3==1.14.0
botocore==1.17.44
Brotli==1.0.9
cbor2==5.4.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dnspython==2.0.0
docutils==0.15.2
ecdsa==0.14.1
email-validator==1.1.2
envs==1.3
et-xmlfile==1.1.0
fancycompleter==0.9.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Compress==1.8.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fsspec==0.8.4
future==0.18.2
idna==2.10
imbalanced-learn==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.0.1
plotly==4.14.1
pyasn1==0.4.8
pycognito==2021.3.1
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrepl==0.9.0
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
s3fs==0.5.1
s3transfer==0.3.3
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wmctrl==0.4
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
